refactor: move utility test files to src/utils/__tests__/
Move test files for moved utilities to be adjacent to their modules,
per CLAUDE.md convention. Renames to match module names:
- IFunctionSymbol.test.ts → FunctionUtils.test.ts
- IParameterInfo.test.ts → ParameterUtils.test.ts
- IScopeSymbol.test.ts → ScopeUtils.test.ts
- TPrimitiveKind.test.ts → PrimitiveKindUtils.test.ts
- TType.test.ts → TTypeUtils.test.ts
- TypeResolver.test.ts → TypeResolver.test.ts
