+++ src/utils/__tests__/FunctionUtils.test.ts
@@ -2,10 +2,10 @@
  * Tests for IFunctionSymbol - C-Next function symbol representation
  */
 import { describe, it, expect } from "vitest";
-import FunctionUtils from "../../../utils/FunctionUtils";
-import ScopeUtils from "../../../utils/ScopeUtils";
-import ParameterUtils from "../../../utils/ParameterUtils";
-import TTypeUtils from "../../../utils/TTypeUtils";
+import FunctionUtils from "../FunctionUtils";
+import ScopeUtils from "../ScopeUtils";
+import ParameterUtils from "../ParameterUtils";
+import TTypeUtils from "../TTypeUtils";
 
 describe("IFunctionSymbol", () => {
   describe("FunctionUtils.create", () => {
